requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
tensorflow==2.8.0
scikit-learn==0.24.2
numpy==1.21.0
pandas==1.3.0
matplotlib==3.4.2
seaborn==0.11.1
opencv-python==4.5.3.56
pillow==8.3.1
joblib==1.0.1
python-dotenv==0.19.0
requests==2.26.0
oauthlib==3.1.1
pyjwt==2.1.0
flask-mail==0.9.1 